Telegram_bot/requirements.txt
# ── Telegram ──────────────────────────────────────────────
python-telegram-bot==20.7

# ── Google Gemini (handles YouTube natively) ──────────────
google-generativeai==0.5.4

# ── Embeddings (FREE local — for RAG Q&A) ─────────────────
sentence-transformers==2.5.1
torch==2.2.2                    # CPU-only is fine

# ── Vector Search ─────────────────────────────────────────
faiss-cpu==1.7.4
numpy==1.26.4

# ── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.1
redis==5.0.1                    # optional caching
